requirements.txt
psycopg2-binary==2.9.9
SQLAlchemy==2.0.29
Flask==3.0.3
Flask-SQLAlchemy==3.1.1